feat: 웹뷰 메시지 타입 분리 및 인증 방식 변경
