+++ src/webview/store/useWebRef.store.ts
@@ -1,7 +1,7 @@
 import React from 'react';
 import WebView from 'react-native-webview';
 import {create} from 'zustand';
-import {PostMessageMethodType} from '../types/message.type';
+import {PostMessageMethodType} from '../types/post-message.type';
 
 interface WebRefStore {
   webRef: React.RefObject<WebView>;
